WIP: Factored out makeWasmModule
