+++ src/core/OptimizerCache.ts
@@ -1,4 +1,4 @@
-import * as t from "../types"
+import * as t from "./types"
 
 const paramSizeSlots = 2
